+++ src/geometry/rotation_conversion.rs
@@ -3,6 +3,9 @@ use num::Zero;
 use alga::general::{Real, SubsetOf, SupersetOf};
 use alga::linear::Rotation as AlgaRotation;
 
+#[cfg(feature = "mint")]
+use mint;
+
 use core::{DefaultAllocator, MatrixN};
 use core::dimension::{DimName, DimNameSum, DimNameAdd, DimMin, U1};
 use core::allocator::Allocator;
@@ -22,8 +25,8 @@ use geometry::{Point, Translation, Rotation, UnitQuaternion, UnitComplex, Isomet
  * Rotation  -> Similarity
  * Rotation  -> Transform
  * Rotation  -> Matrix (homogeneous)
- * mint::EulerAngles -> RotationBase
- 
+ * mint::EulerAngles -> Rotation
+
 */
 
 
@@ -205,11 +208,7 @@ impl<N1, N2, D> SubsetOf<MatrixN<N2, DimNameSum<D, U1>>> for Rotation<N1, D>
 }
 
 #[cfg(feature = "mint")]
-impl<N, S> From<mint::EulerAngles<N, mint::IntraXYZ>> for  RotationBase<N, U3, S>
-where N: Real,
-      S: OwnedStorage<N, U3, U3>,
-      S::Alloc: OwnedAllocator<N, U3, U3, S> {
-
+impl<N: Real> From<mint::EulerAngles<N, mint::IntraXYZ>> for Rotation3<N> {
     fn from(euler: mint::EulerAngles<N, mint::IntraXYZ>) -> Self {
         Self::from_euler_angles(euler.a, euler.b, euler.c)
     }
